refactor(interfaces): Use typeof ... === 'function' to test for methods
Testing for

    'name' in object

or

    obj.name !== undefined

only checks for the existence of the property (and in the latter
case that the property is not set to undefined).  That's fine if
the interface specifies a property of indeterminate type, but in
the usual case that the interface member is a method we can do
one better and check to make sure the property's value is
callable.