refactor(shortcuts): Rename import isDeletable -> isIDeletable etc.
Some of the existing code is confusing to read because e.g.
isDeletable doesn't check if an item .isDeletable(), but only
whether it is an IDeletable.

By renaming these imports the shortcut precondition functions are
easier to understand, and allows a subsequent to commit to add
an isCopyable function that actually checks copyability.