fix: A bunch of focus issues.
This removes Flyout's ability to be tab-focused and insteads relies on
treating it as a workspace, removing the extra tab stop.

This fixes an issue with FocusManager getting semi-permanently out of
sync with the DOM.

This reintroduces automatic tab management for nodes, and fixes an issue
with auto tab management for trees that would break tabbing order.